Change upload URI format to https://bucket-name.s3.amazon.aws.com
This code was not working out of the box, at least in my region, eu-west-1.  The above code change allows the upload to go through and keeps the environmental variables the same. An alternative fix would be to specify the upload URI differently and change the code the other way. I've chosen this one as it's less redundant because it doesn't repeat the bucket name twice in the env variables.